Fix expired jobs showing in browse results
